move service worker
